Implementation of Logistic Regression via Perceptron Trick
Perceptron Trick is a mechanism via which we start with a random hyperplane ( by taking taking desired coefficients for our features ) in space and get to the best fit hyperplane which segregates categories of output features in the best possible way it can. To do this, we continuously take datapoints from our dataset, analyze it and then apply the necessary transformation that is required to make our line better with respect to that point.